allensdk/brain_observatory/behavior/data_objects/cell_specimens/cell_specimens.py
from typing import Optional, Tuple

import numpy as np
import pandas as pd
from pynwb import NWBFile, ProcessingModule
from pynwb.ophys import OpticalChannel, ImageSegmentation

import allensdk.brain_observatory.roi_masks as roi
from allensdk.brain_observatory.behavior.data_files.demix_file import DemixFile
from allensdk.brain_observatory.behavior.data_files.dff_file import DFFFile
from allensdk.brain_observatory.behavior.data_files.event_detection_file \
    import \
    EventDetectionFile
from allensdk.brain_observatory.behavior.data_objects import DataObject
from allensdk.brain_observatory.behavior.data_objects.base \
    .readable_interfaces import \
    JsonReadableInterface, LimsReadableInterface, NwbReadableInterface
from allensdk.brain_observatory.behavior.data_objects.base \
    .writable_interfaces import \
    NwbWritableInterface
from allensdk.brain_observatory.behavior.data_objects.cell_specimens.events \
    import \
    Events
from allensdk.brain_observatory.behavior.data_objects.cell_specimens.traces \
    .corrected_fluorescence_traces import \
    CorrectedFluorescenceTraces
from allensdk.brain_observatory.behavior.data_objects.cell_specimens.traces \
    .dff_traces import \
    DFFTraces
from allensdk.brain_observatory.behavior.data_objects.metadata \
    .ophys_experiment_metadata.field_of_view_shape import \
    FieldOfViewShape
from allensdk.brain_observatory.behavior.data_objects.metadata \
    .ophys_experiment_metadata.imaging_plane import \
    ImagingPlane
from allensdk.brain_observatory.behavior.data_objects.timestamps \
    .ophys_timestamps import \
    OphysTimestamps
from allensdk.brain_observatory.behavior.image_api import Image
from allensdk.brain_observatory.nwb import CELL_SPECIMEN_COL_DESCRIPTIONS
from allensdk.brain_observatory.nwb.nwb_utils import add_image_to_nwb
from allensdk.internal.api import PostgresQueryMixin


class EventsParams:
    """Container for arguments to event detection"""

    def __init__(self,
                 filter_scale: float = 2,
                 filter_n_time_steps: int = 20):
        """
        :param filter_scale
            See Events.filter_scale
        :param filter_n_time_steps
            See Events.filter_n_time_steps
        """
        self._filter_scale = filter_scale
        self._filter_n_time_steps = filter_n_time_steps

    @property
    def filter_scale(self):
        return self._filter_scale

    @property
    def filter_n_time_steps(self):
        return self._filter_n_time_steps


class CellSpecimenMeta(DataObject, LimsReadableInterface,
                       JsonReadableInterface, NwbReadableInterface):
    """Cell specimen metadata"""
    def __init__(self, imaging_plane: ImagingPlane, emission_lambda=520.0):
        super().__init__(name='cell_spcimen_meta', value=self)
        self._emission_lambda = emission_lambda
        self._imaging_plane = imaging_plane

    @property
    def emission_lambda(self):
        return self._emission_lambda

    @property
    def imaging_plane(self):
        return self._imaging_plane

    @classmethod
    def from_lims(cls, ophys_experiment_id: int,
                  lims_db: PostgresQueryMixin,
                  ophys_timestamps: OphysTimestamps) -> "CellSpecimenMeta":
        imaging_plane_meta = ImagingPlane.from_lims(
            ophys_experiment_id=ophys_experiment_id, lims_db=lims_db,
            ophys_timestamps=ophys_timestamps)
        return cls(imaging_plane=imaging_plane_meta)

    @classmethod
    def from_json(cls, dict_repr: dict,
                  ophys_timestamps: OphysTimestamps) -> "CellSpecimenMeta":
        imaging_plane_meta = ImagingPlane.from_json(
            dict_repr=dict_repr, ophys_timestamps=ophys_timestamps)
        return cls(imaging_plane=imaging_plane_meta)

    @classmethod
    def from_nwb(cls, nwbfile: NWBFile) -> "CellSpecimenMeta":
        ophys_module = nwbfile.processing['ophys']
        image_seg = ophys_module.data_interfaces['image_segmentation']
        plane_segmentations = image_seg.plane_segmentations
        cell_specimen_table = plane_segmentations['cell_specimen_table']

        imaging_plane = cell_specimen_table.imaging_plane
        optical_channel = imaging_plane.optical_channel[0]
        emission_lambda = optical_channel.emission_lambda

        imaging_plane = ImagingPlane.from_nwb(nwbfile=nwbfile)
        return CellSpecimenMeta(emission_lambda=emission_lambda,
                                imaging_plane=imaging_plane)


class CellSpecimens(DataObject, LimsReadableInterface,
                    JsonReadableInterface, NwbReadableInterface,
                    NwbWritableInterface):
    def __init__(self,
                 cell_specimen_table: pd.DataFrame,
                 meta: CellSpecimenMeta,
                 dff_traces: DFFTraces,
                 corrected_fluorescence_traces: CorrectedFluorescenceTraces,
                 events: Events,
                 ophys_timestamps: OphysTimestamps,
                 segmentation_mask_image_spacing: Tuple,
                 exclude_invalid_rois=True):
        """
        A container for cell specimens including traces, events, metadata, etc.

        Parameters
        ----------
        cell_specimen_table
            index cell_specimen_id
            columns:
                - cell_roi_id
                - height
                - mask_image_plane
                - max_correction_down
                - max_correction_left
                - max_correction_right
                - max_correction_up
                - roi_mask
                - valid_roi
                - width
                - x
                - y
        meta
        dff_traces
        corrected_fluorescence_traces
        events
        ophys_timestamps
        segmentation_mask_image_spacing
            Spacing to pass to sitk when constructing segmentation mask image
        exclude_invalid_rois
            Whether to exclude invalid rois

        """
        super().__init__(name='cell_specimen_table', value=self)

        # Validate ophys timestamps, traces
        ophys_timestamps = ophys_timestamps.validate(
            number_of_frames=dff_traces.get_number_of_frames())
        self._validate_traces(
            ophys_timestamps=ophys_timestamps, dff_traces=dff_traces,
            corrected_fluorescence_traces=corrected_fluorescence_traces,
            cell_roi_ids=cell_specimen_table['cell_roi_id'].values)

        if exclude_invalid_rois:
            cell_specimen_table = cell_specimen_table[
                cell_specimen_table['valid_roi']]

        # Filter/reorder rois according to cell_specimen_table
        dff_traces.filter_and_reorder(
            roi_ids=cell_specimen_table['cell_roi_id'].values)
        corrected_fluorescence_traces.filter_and_reorder(
            roi_ids=cell_specimen_table['cell_roi_id'].values)

        # Note: setting raise_if_rois_missing to False for events, since
        # there seem to be cases where cell_specimen_table contains rois not in
        # events
        # See ie https://app.zenhub.com/workspaces/allensdk-10-5c17f74db59cfb36f158db8c/issues/alleninstitute/allensdk/2139     # noqa
        events.filter_and_reorder(
            roi_ids=cell_specimen_table['cell_roi_id'].values,
            raise_if_rois_missing=False)

        self._meta = meta
        self._cell_specimen_table = cell_specimen_table
        self._dff_traces = dff_traces
        self._corrected_fluorescence_traces = corrected_fluorescence_traces
        self._events = events
        self._segmentation_mask_image = self._get_segmentation_mask_image(
            spacing=segmentation_mask_image_spacing)

    @property
    def table(self) -> pd.DataFrame:
        return self._cell_specimen_table

    @property
    def roi_masks(self) -> pd.DataFrame:
        return self._cell_specimen_table[['cell_roi_id', 'roi_mask']]

    @property
    def meta(self) -> CellSpecimenMeta:
        return self._meta

    @property
    def dff_traces(self) -> pd.DataFrame:
        df = self.table[['cell_roi_id']].join(self._dff_traces.value,
                                              on='cell_roi_id')
        return df

    @property
    def corrected_fluorescence_traces(self) -> pd.DataFrame:
        df = self.table[['cell_roi_id']].join(
            self._corrected_fluorescence_traces.value, on='cell_roi_id')
        return df

    @property
    def events(self) -> pd.DataFrame:
        df = self.table.reset_index()
        df = df[['cell_roi_id', 'cell_specimen_id']] \
            .merge(self._events.value, on='cell_roi_id')
        df = df.set_index('cell_specimen_id')
        return df

    @property
    def segmentation_mask_image(self) -> Image:
        return self._segmentation_mask_image

    @classmethod
    def from_lims(cls, ophys_experiment_id: int,
                  lims_db: PostgresQueryMixin,
                  ophys_timestamps: OphysTimestamps,
                  segmentation_mask_image_spacing: Tuple,
                  exclude_invalid_rois=True,
                  events_params: Optional[EventsParams] = None) \
            -> "CellSpecimens":
        def _get_ophys_cell_segmentation_run_id() -> int:
            """Get the ophys cell segmentation run id associated with an
            ophys experiment id"""
            query = """
                    SELECT oseg.id
                    FROM ophys_experiments oe
                    JOIN ophys_cell_segmentation_runs oseg
                    ON oe.id = oseg.ophys_experiment_id
                    WHERE oseg.current = 't'
                    AND oe.id = {};
                    """.format(ophys_experiment_id)
            return lims_db.fetchone(query, strict=True)

        def _get_cell_specimen_table():
            ophys_cell_seg_run_id = _get_ophys_cell_segmentation_run_id()
            query = """
                    SELECT *
                    FROM cell_rois cr
                    WHERE cr.ophys_cell_segmentation_run_id = {};
                    """.format(ophys_cell_seg_run_id)
            initial_cs_table = pd.read_sql(query, lims_db.get_connection())
            cst = initial_cs_table.rename(
                columns={'id': 'cell_roi_id', 'mask_matrix': 'roi_mask'})
            cst.drop(['ophys_experiment_id',
                      'ophys_cell_segmentation_run_id'],
                     inplace=True, axis=1)
            cst = cst.to_dict()
            fov_shape = FieldOfViewShape.from_lims(
                ophys_experiment_id=ophys_experiment_id, lims_db=lims_db)
            cst = cls._postprocess(
                cell_specimen_table=cst, fov_shape=fov_shape)
            return cst

        def _get_dff_traces():
            dff_file = DFFFile.from_lims(
                ophys_experiment_id=ophys_experiment_id,
                db=lims_db)
            return DFFTraces.from_data_file(
                dff_file=dff_file)

        def _get_corrected_fluorescence_traces():
            demix_file = DemixFile.from_lims(
                ophys_experiment_id=ophys_experiment_id,
                db=lims_db)
            return CorrectedFluorescenceTraces.from_data_file(
                demix_file=demix_file)

        def _get_events():
            events_file = EventDetectionFile.from_lims(
                ophys_experiment_id=ophys_experiment_id,
                db=lims_db)
            return cls._get_events(events_file=events_file,
                                   events_params=events_params)

        cell_specimen_table = _get_cell_specimen_table()
        meta = CellSpecimenMeta.from_lims(
            ophys_experiment_id=ophys_experiment_id, lims_db=lims_db,
            ophys_timestamps=ophys_timestamps)
        dff_traces = _get_dff_traces()
        corrected_fluorescence_traces = _get_corrected_fluorescence_traces()
        events = _get_events()

        return CellSpecimens(
            cell_specimen_table=cell_specimen_table, meta=meta,
            dff_traces=dff_traces,
            corrected_fluorescence_traces=corrected_fluorescence_traces,
            events=events,
            ophys_timestamps=ophys_timestamps,
            segmentation_mask_image_spacing=segmentation_mask_image_spacing,
            exclude_invalid_rois=exclude_invalid_rois
        )

    @classmethod
    def from_json(cls, dict_repr: dict,
                  ophys_timestamps: OphysTimestamps,
                  segmentation_mask_image_spacing: Tuple,
                  exclude_invalid_rois=True,
                  events_params: Optional[EventsParams] = None) \
            -> "CellSpecimens":
        cell_specimen_table = dict_repr['cell_specimen_table_dict']
        fov_shape = FieldOfViewShape.from_json(dict_repr=dict_repr)
        cell_specimen_table = cls._postprocess(
            cell_specimen_table=cell_specimen_table, fov_shape=fov_shape)

        def _get_dff_traces():
            dff_file = DFFFile.from_json(dict_repr=dict_repr)
            return DFFTraces.from_data_file(
                dff_file=dff_file)

        def _get_corrected_fluorescence_traces():
            demix_file = DemixFile.from_json(dict_repr=dict_repr)
            return CorrectedFluorescenceTraces.from_data_file(
                demix_file=demix_file)

        def _get_events():
            events_file = EventDetectionFile.from_json(dict_repr=dict_repr)
            return cls._get_events(events_file=events_file,
                                   events_params=events_params)

        meta = CellSpecimenMeta.from_json(dict_repr=dict_repr,
                                          ophys_timestamps=ophys_timestamps)
        dff_traces = _get_dff_traces()
        corrected_fluorescence_traces = _get_corrected_fluorescence_traces()
        events = _get_events()
        return CellSpecimens(
            cell_specimen_table=cell_specimen_table, meta=meta,
            dff_traces=dff_traces,
            corrected_fluorescence_traces=corrected_fluorescence_traces,
            events=events,
            ophys_timestamps=ophys_timestamps,
            segmentation_mask_image_spacing=segmentation_mask_image_spacing,
            exclude_invalid_rois=exclude_invalid_rois)

    @classmethod
    def from_nwb(cls, nwbfile: NWBFile,
                 segmentation_mask_image_spacing: Tuple,
                 exclude_invalid_rois=True,
                 events_params: Optional[EventsParams] = None) \
            -> "CellSpecimens":
        # NOTE: ROI masks are stored in full frame width and height arrays
        ophys_module = nwbfile.processing['ophys']
        image_seg = ophys_module.data_interfaces['image_segmentation']
        plane_segmentations = image_seg.plane_segmentations
        cell_specimen_table = plane_segmentations['cell_specimen_table']

        def _read_table(cell_specimen_table):
            df = cell_specimen_table.to_dataframe()

            # Ensure int64 used instead of int32
            df = df.astype(
                {col: 'int64' for col in df.select_dtypes('int32').columns})

            # Because pynwb stores this field as "image_mask", it is renamed
            # here
            df = df.rename(columns={'image_mask': 'roi_mask'})

            df.index.rename('cell_roi_id', inplace=True)
            df['cell_specimen_id'] = [None if id_ == -1 else id_
                                      for id_ in df['cell_specimen_id'].values]

            df.reset_index(inplace=True)
            df.set_index('cell_specimen_id', inplace=True)
            return df

        df = _read_table(cell_specimen_table=cell_specimen_table)
        meta = CellSpecimenMeta.from_nwb(nwbfile=nwbfile)
        dff_traces = DFFTraces.from_nwb(nwbfile=nwbfile)
        corrected_fluorescence_traces = CorrectedFluorescenceTraces.from_nwb(
            nwbfile=nwbfile)

        def _get_events():
            ep = EventsParams() if events_params is None else events_params
            return Events.from_nwb(
                nwbfile=nwbfile, filter_scale=ep.filter_scale,
                filter_n_time_steps=ep.filter_n_time_steps)

        events = _get_events()
        ophys_timestamps = OphysTimestamps.from_nwb(nwbfile=nwbfile)

        return CellSpecimens(
            cell_specimen_table=df, meta=meta, dff_traces=dff_traces,
            corrected_fluorescence_traces=corrected_fluorescence_traces,
            events=events,
            ophys_timestamps=ophys_timestamps,
            segmentation_mask_image_spacing=segmentation_mask_image_spacing,
            exclude_invalid_rois=exclude_invalid_rois)

    def to_nwb(self, nwbfile: NWBFile,
               ophys_timestamps: OphysTimestamps) -> NWBFile:
        """
        :param nwbfile
            In-memory nwb file object
        :param ophys_timestamps
            ophys timestamps
        """
        # 1. Add cell specimen table
        cell_roi_table = self.table.reset_index().set_index(
            'cell_roi_id')
        metadata = nwbfile.lab_meta_data['metadata']

        device = nwbfile.get_device()

        # FOV:
        fov_width = metadata.field_of_view_width
        fov_height = metadata.field_of_view_height
        imaging_plane_description = \
            "{} field of view in {} at depth {} " \
            "um".format(
                (fov_width, fov_height),
                self._meta.imaging_plane.targeted_structure,
                metadata.imaging_depth)

        # Optical Channel:
        optical_channel = OpticalChannel(
            name='channel_1',
            description='2P Optical Channel',
            emission_lambda=self._meta.emission_lambda)

        # Imaging Plane:
        imaging_plane = nwbfile.create_imaging_plane(
            name='imaging_plane_1',
            optical_channel=optical_channel,
            description=imaging_plane_description,
            device=device,
            excitation_lambda=self._meta.imaging_plane.excitation_lambda,
            imaging_rate=self._meta.imaging_plane.ophys_frame_rate,
            indicator=self._meta.imaging_plane.indicator,
            location=self._meta.imaging_plane.targeted_structure)

        # Image Segmentation:
        image_segmentation = ImageSegmentation(name="image_segmentation")

        if 'ophys' not in nwbfile.processing:
            ophys_module = ProcessingModule('ophys', 'Ophys processing module')
            nwbfile.add_processing_module(ophys_module)
        else:
            ophys_module = nwbfile.processing['ophys']

        ophys_module.add_data_interface(image_segmentation)

        # Plane Segmentation:
        plane_segmentation = image_segmentation.create_plane_segmentation(
            name='cell_specimen_table',
            description="Segmented rois",
            imaging_plane=imaging_plane)

        for col_name in cell_roi_table.columns:
            # the columns 'roi_mask', 'pixel_mask', and 'voxel_mask' are
            # already defined in the nwb.ophys::PlaneSegmentation Object
            if col_name not in ['id', 'mask_matrix', 'roi_mask',
                                'pixel_mask', 'voxel_mask']:
                # This builds the columns with name of column and description
                # of column both equal to the column name in the cell_roi_table
                plane_segmentation.add_column(
                    col_name,
                    CELL_SPECIMEN_COL_DESCRIPTIONS.get(
                        col_name,
                        "No Description Available"))

        # go through each roi and add it to the plan segmentation object
        for cell_roi_id, table_row in cell_roi_table.iterrows():
            # NOTE: The 'roi_mask' in this cell_roi_table has already been
            # processing by the function from
            # allensdk.brain_observatory.behavior.session_apis.data_io
            # .ophys_lims_api
            # get_cell_specimen_table() method. As a result, the ROI is
            # stored in
            # an array that is the same shape as the FULL field of view of the
            # experiment (e.g. 512 x 512).
            mask = table_row.pop('roi_mask')

            csid = table_row.pop('cell_specimen_id')
            table_row['cell_specimen_id'] = -1 if csid is None else csid
            table_row['id'] = cell_roi_id
            plane_segmentation.add_roi(image_mask=mask, **table_row.to_dict())

        # 2. Add DFF traces
        self._dff_traces.to_nwb(nwbfile=nwbfile,
                                ophys_timestamps=ophys_timestamps)

        # 3. Add Corrected fluorescence traces
        self._corrected_fluorescence_traces.to_nwb(nwbfile=nwbfile)

        # 4. Add events
        self._events.to_nwb(nwbfile=nwbfile)

        # 5. Add segmentation mask image
        add_image_to_nwb(nwbfile=nwbfile,
                         image_data=self._segmentation_mask_image,
                         image_name='segmentation_mask_image')

        return nwbfile

    def _get_segmentation_mask_image(self, spacing: tuple) -> Image:
        """a 2D binary image of all cell masks

        Parameters
        ----------
        spacing
            See image_api.Image for details

        Returns
        ----------
        allensdk.brain_observatory.behavior.image_api.Image:
            array-like interface to segmentation_mask image data and
            metadata
        """
        mask_data = np.sum(self.roi_masks['roi_mask']).astype(int)

        mask_image = Image(
            data=mask_data,
            spacing=spacing,
            unit='mm'
        )
        return mask_image

    @staticmethod
    def _postprocess(cell_specimen_table: dict,
                     fov_shape: FieldOfViewShape) -> pd.DataFrame:
        """Converts raw cell_specimen_table dict to dataframe"""
        cell_specimen_table = pd.DataFrame.from_dict(
            cell_specimen_table).set_index(
            'cell_roi_id').sort_index()
        fov_width = fov_shape.width
        fov_height = fov_shape.height

        # Convert cropped ROI masks to uncropped versions
        roi_mask_list = []
        for cell_roi_id, table_row in cell_specimen_table.iterrows():
            # Deserialize roi data into AllenSDK RoiMask object
            curr_roi = roi.RoiMask(image_w=fov_width, image_h=fov_height,
                                   label=None, mask_group=-1)
            curr_roi.x = table_row['x']
            curr_roi.y = table_row['y']
            curr_roi.width = table_row['width']
            curr_roi.height = table_row['height']
            curr_roi.mask = np.array(table_row['roi_mask'])
            roi_mask_list.append(curr_roi.get_mask_plane().astype(np.bool))

        cell_specimen_table['roi_mask'] = roi_mask_list
        cell_specimen_table = cell_specimen_table[
            sorted(cell_specimen_table.columns)]

        cell_specimen_table.index.rename('cell_roi_id', inplace=True)
        cell_specimen_table.reset_index(inplace=True)
        cell_specimen_table.set_index('cell_specimen_id', inplace=True)
        return cell_specimen_table

    def _validate_traces(
            self, ophys_timestamps: OphysTimestamps,
            dff_traces: DFFTraces,
            corrected_fluorescence_traces: CorrectedFluorescenceTraces,
            cell_roi_ids: np.ndarray):
        """validates traces"""
        trace_col_map = {
            'dff_traces': 'dff',
            'corrected_fluorescence_traces': 'corrected_fluorescence'
        }
        for traces in (dff_traces, corrected_fluorescence_traces):
            # validate traces contain expected roi ids
            if not np.in1d(traces.value.index, cell_roi_ids).all():
                raise RuntimeError(f"{traces.name} contains ROI IDs that "
                                   f"are not in "
                                   f"cell_specimen_table.cell_roi_id")
            if not np.in1d(cell_roi_ids, traces.value.index).all():
                raise RuntimeError(f"cell_specimen_table contains ROI IDs "
                                   f"that are not in {traces.name}")

            # validate traces contain expected timepoints
            num_trace_timepoints = len(traces.value.iloc[0]
                                       [trace_col_map[traces.name]])
            num_ophys_timestamps = ophys_timestamps.value.shape[0]
            if num_trace_timepoints != num_ophys_timestamps:
                raise RuntimeError(f'{traces.name} contains '
                                   f'{num_trace_timepoints} '
                                   f'but there are {num_ophys_timestamps} '
                                   f'ophys timestamps')

    @staticmethod
    def _get_events(events_file: EventDetectionFile,
                    events_params: Optional[EventsParams] = None):
        if events_params is None:
            events_params = EventsParams()
        return Events.from_data_file(
            events_file=events_file,
            filter_scale=events_params.filter_scale,
            filter_n_time_steps=events_params.filter_n_time_steps)
